Defer SoapClient creation
For a couple reasons it is ideal to defer the creation of the SoapClient
until it is going to be used. In some cases it can lead to non-trivial
performance improvements.

1) SoapClient validates the WSDL during construction. When using using
the services you may accept the client/service as a dependency but only use it
occasionaly(caching etc). You will still take the performance hit even
when not used which should not happen in an IOC model.
2) Constructing SoapClient makes a web request which can throw a
SoapFault during construction. This is a very bad thing in general. It
can cause very frustrating failures outside try catch structures.
3) If you don't set the host there is an additional soap request during
construction even when the service isn't used.